Item10188: PleaseFeelFreeToModifyOrReWrite
Comitting this broken, not-finished-re-re-re-factoring. Maybe it will prompt
someone to do something between now and whenever I finish it.

In an ideal world we wouldn't need this, just use git submodules and forget
about svn. But pseudo-install isn't quite smart enough, we haven't figured out
TasksContrib integration. Also want to move the repo-awareness code out of
pseudo-install into a reusable form here.

The intention for this is to get some experience for non-core stuff, which
might inform us on how to proceed.

It's very important not to fragment our community by alienating those who don't
want to use git yet. IMHO this means a gradual phase-in. Maybe we can do a
"moved to git" thing for svn so that new commits aren't made on plugins whose
authors have moved to git; maybe there's even some hack we can use to keep svn
up-to-date after such a migration.

My apologies if you look at this and then wondered why I bothered ;-) It was
holidays and I'd been re-reading Damian Conway's PBP and trying out the very
nice Getopts::Euclid, too bad it's so buggy...

git-svn-id: http://svn.foswiki.org/trunk/FoswikiGitDevPlugin@10501 0b4bb1d4-4e5a-0410-9cc4-b2b747904278